+++ extensions/ai-gen/src/components/ImageActions.tsx
@@ -9,8 +9,9 @@ import downloadTempFile from "../lib/downloadTempFile";
 import copyFileToClipboard from "../lib/copyFileToClipboard";
 
 type ImageActionProps = {
+  prompt?: string;
+  image?: string;
   url: string;
-  prompt: string;
   size: CreateImageRequestSizeEnum;
   n: string;
   variationCount: number;
@@ -18,13 +19,13 @@ type ImageActionProps = {
 };
 
 export function ImageActions(props: ImageActionProps) {
-  const { url, prompt, n, size, variationCount, showDetailAction } = props;
+  const { url, prompt, image, n, size, variationCount, showDetailAction } = props;
   const number = parseInt(n, 10);
 
   const { push } = useNavigation();
   async function createVariationAction(url: string, count: number) {
-    const file = await downloadTempFile(url);
-    push(<ImagesGrid prompt={prompt} file={file} n={n} size={size} variationCount={count + 1} />);
+    const file = url.startsWith("http") ? await downloadTempFile(url) : url;
+    push(<ImagesGrid prompt={prompt} image={file} n={n} size={size} variationCount={count + 1} />);
   }
 
   return (
@@ -50,7 +51,7 @@ export function ImageActions(props: ImageActionProps) {
           <Action.Push
             title="View Details"
             icon={Icon.Eye}
-            target={<ImageDetails url={url} opt={{ prompt, n: number, size: props.size, variationCount }} />}
+            target={<ImageDetails url={url} opt={{ prompt, n: number, size: props.size, variationCount, image }} />}
             shortcut={{ modifiers: ["cmd", "shift"], key: "d" }}
           />
         )}