+++ extensions/ai-gen/src/components/ImageDetails.tsx
@@ -1,3 +1,6 @@
+import path from "path";
+import { URL } from "url";
+
 import { Detail } from "@raycast/api";
 
 import { CreateVariationRequest } from "../hooks/useOpenAIImageApi";
@@ -5,9 +8,13 @@ import { ImageActions } from "./ImageActions";
 
 export function ImageDetails(props: {
   url: string;
-  opt: CreateVariationRequest & { prompt: string; variationCount?: number };
+  opt: CreateVariationRequest & { prompt?: string; variationCount?: number; image?: string };
 }) {
   const { url, opt } = props;
+  let fileName = "";
+  if (opt.image) {
+    fileName = path.basename(opt.image);
+  }
 
   return (
     <Detail
@@ -17,6 +24,7 @@ export function ImageDetails(props: {
           showDetailAction={false}
           url={url}
           prompt={opt.prompt}
+          image={opt.image}
           size={opt.size}
           n={opt.n.toString()}
           variationCount={opt.variationCount ?? 0}
@@ -24,7 +32,11 @@ export function ImageDetails(props: {
       }
       metadata={
         <Detail.Metadata>
-          <Detail.Metadata.Label title="Prompt" text={opt.prompt} />
+          {opt.prompt ? (
+            <Detail.Metadata.Label title="Prompt" text={opt.prompt} />
+          ) : fileName && opt.image ? (
+            <Detail.Metadata.Link title="Original Image" text={fileName} target={"file:///" + opt.image} />
+          ) : null}
           <Detail.Metadata.Label title="Size" text={opt.size} />
           <Detail.Metadata.Label
             title="Variation"