+++ extensions/ai-gen/CHANGELOG.md
@@ -1,4 +1,7 @@
-# ai-gen Changelog
+# OpenAI Generator Changelog
+
+## [Add Create Image Variation command] - 2022-11-10
+- Adds a new Create Image Variation command to start a variation from an image on your local computer
 
 ## [Advanced text completion fixes] - 2022-11-09
 - Fix text completion advanced options not saving if you hide the UI